Added alternate logging and multi-file/directory input!
Added an option for alternate logging (-al) that includes the file name,
total number of departures and arrivals, and the file size for file size
analytics.  Also included multi-file or multi-directory input.  If a
directory is given, it will run on all .h264 files in the directory.
You can include multiple files and directories in a single call of the
utility!